Rework event handlers for partial classes
- Moved methods from shared classes to specific technologies becuase of unavailable APIs.
