+++ appveyor.yml
@@ -121,6 +121,8 @@
   skip_tags: false
 
   test: off
+
+  shallow_clone: false
   
   pull_requests:
     do_not_increment_build_number: true
