+++ source/nanoFramework.Tools.DebugLibrary.Shared/PortSerial/EventHandlerForSerialDevice.cs
@@ -200,118 +200,6 @@ namespace nanoFramework.Tools.Debugger.Serial
             }
         }
 
-        /// <summary>
-        /// This method opens the device using the WinRT Serial API. After the device is opened, save the device
-        /// so that it can be used across scenarios.
-        ///
-        /// It is important that the FromIdAsync call is made on the UI thread because the consent prompt can only be displayed
-        /// on the UI thread.
-        /// 
-        /// This method is used to reopen the device after the device reconnects to the computer and when the app resumes.
-        /// </summary>
-        /// <param name="deviceInfo">Device information of the device to be opened</param>
-        /// <param name="deviceSelector">The AQS used to find this device</param>
-        /// <returns>True if the device was successfully opened, false if the device could not be opened for well known reasons.
-        /// An exception may be thrown if the device could not be opened for extraordinary reasons.</returns>
-        public async Task<bool> OpenDeviceAsync(DeviceInformation deviceInfo, string deviceSelector)
-        {
-            await Task.Delay(250);
-
-            device = await SerialDevice.FromIdAsync(deviceInfo.Id);
-
-            bool successfullyOpenedDevice = false;
-
-            try
-            {
-                // Device could have been blocked by user or the device has already been opened by another app.
-                if (device != null)
-                {
-                    successfullyOpenedDevice = true;
-
-                    deviceInformation = deviceInfo;
-                    this.deviceSelector = deviceSelector;
-
-                    Debug.WriteLine($"Device {deviceInformation.Id} opened");
-
-                    // adjust settings for serial port
-                    device.BaudRate = 115200;
-                    
-                    /////////////////////////////////////////////////////////////
-                    // need to FORCE the parity setting to _NONE_ because        
-                    // the default on the current ST Link is different causing 
-                    // the communication to fail
-                    /////////////////////////////////////////////////////////////
-                    device.Parity = SerialParity.None;
-
-                    device.WriteTimeout = TimeSpan.FromMilliseconds(1000);
-                    device.ReadTimeout = TimeSpan.FromMilliseconds(1000);
-                    device.ErrorReceived += Device_ErrorReceived;
-
-                    // Notify registered callback handle that the device has been opened
-                    deviceConnectedCallback?.Invoke(this, deviceInformation);
-
-                    // Background tasks are not part of the app, so app events will not have an affect on the device
-                    if (!isBackgroundTask && (appSuspendEventHandler == null || appResumeEventHandler == null))
-                    {
-                        RegisterForAppEvents();
-                    }
-
-                    // User can block the device after it has been opened in the Settings charm. We can detect this by registering for the 
-                    // DeviceAccessInformation.AccessChanged event
-                    if (deviceAccessEventHandler == null)
-                    {
-                        RegisterForDeviceAccessStatusChange();
-                    }
-
-                    // Create and register device watcher events for the device to be opened unless we're reopening the device
-                    if (deviceWatcher == null)
-                    {
-                        deviceWatcher = DeviceInformation.CreateWatcher(deviceSelector);
-
-                        RegisterForDeviceWatcherEvents();
-                    }
-
-                    if (!watcherStarted)
-                    {
-                        // Start the device watcher after we made sure that the device is opened.
-                        StartDeviceWatcher();
-                    }
-                }
-                else
-                {
-                    successfullyOpenedDevice = false;
-
-                    //notificationStatus = NotifyType.ErrorMessage;
-
-                    var deviceAccessStatus = DeviceAccessInformation.CreateFromId(deviceInfo.Id).CurrentStatus;
-
-                    switch (deviceAccessStatus)
-                    {
-                        case DeviceAccessStatus.DeniedByUser:
-                            Debug.WriteLine($"Access to the device was blocked by the user : {deviceInfo.Id}");
-                            break;
-
-                        case DeviceAccessStatus.DeniedBySystem:
-                            // This status is most likely caused by app permissions (did not declare the device in the app's package.appxmanifest)
-                            // This status does not cover the case where the device is already opened by another app.
-                            Debug.WriteLine($"Access to the device was blocked by the system : {deviceInfo.Id}");
-                            break;
-
-                        default:
-                            // Most likely the device is opened by another app, but cannot be sure
-                            Debug.WriteLine($"Unknown error, possibly opened by another app : {deviceInfo.Id}");
-                            break;
-                    }
-                }
-            }
-            // catch all because the device open might fail for a number of reasons
-            catch(Exception ex)
-            {
-            }
-
-            return successfullyOpenedDevice;
-        }
-
         private void Device_ErrorReceived(SerialDevice sender, ErrorReceivedEventArgs args)
         {
             throw new NotImplementedException();
@@ -496,23 +384,6 @@ namespace nanoFramework.Tools.Debugger.Serial
         private void OnDeviceAdded(DeviceWatcher sender, DeviceInformation deviceInfo)
         {
             if ((deviceInformation != null) && (deviceInfo.Id == deviceInformation.Id) && !IsDeviceConnected && isEnabledAutoReconnect)
-            {
-            }
-        }
-
-        /// <summary>
-        /// Close the device if the device access was denied by anyone (system or the user) and reopen it if permissions are allowed again
-        /// </summary>
-        /// <param name="sender"></param>
-        /// <param name="eventArgs"></param>
-        private void OnDeviceAccessChanged(DeviceAccessInformation sender, DeviceAccessChangedEventArgs eventArgs)
-        {
-            if ((eventArgs.Status == DeviceAccessStatus.DeniedBySystem)
-                || (eventArgs.Status == DeviceAccessStatus.DeniedByUser))
-            {
-                CloseCurrentlyConnectedDevice();
-            }
-            else if ((eventArgs.Status == DeviceAccessStatus.Allowed) && (deviceInformation != null) && isEnabledAutoReconnect)
             {
             }
         }
