+++ source/nanoFramework.Tools.DebugLibrary.UWP/nanoFramework.Tools.DebugLibrary.UWP.csproj
@@ -129,6 +129,7 @@ For .NET look for the respective Nuget package.</PackageReleaseNotes>
     <Compile Include="PortDefinitions\PortBase.cs" />
     <Compile Include="PortSerial\EventHandlerForSerialDevice.cs" />
     <Compile Include="PortSerial\SerialPort.cs" />
+    <Compile Include="PortUsb\EventHandlerForUsbDevice.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Properties\nanoFramework.Tools.DebugLibrary.UWP.rd.xml" />
   </ItemGroup>
